Refactor `source_instructions` into a multi file mod `instructions`
