+++ probe-rs/src/debug/instructions.rs
@@ -1,0 +1,68 @@
+mod block;
+mod breakpoint;
+mod instruction;
+mod sequence;
+use instruction::Instruction;
+
+use super::{
+    unit_info::{self},
+    ColumnType, DebugInfo,
+};
+pub use breakpoint::VerifiedBreakpoint;
+use typed_path::TypedPathBuf;
+
+pub(crate) fn serialize_typed_path<S>(
+    path: &Option<TypedPathBuf>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: serde::Serializer,
+{
+    match path {
+        Some(path) => serializer.serialize_str(&path.to_string_lossy()),
+        None => serializer.serialize_none(),
+    }
+}
+
+/// A specific location in source code.
+/// Each unique line, column, file and directory combination is a unique source location.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize)]
+pub struct SourceLocation {
+    /// The line number in the source file with zero based indexing.
+    pub line: Option<u64>,
+    /// The column number in the source file with zero based indexing.
+    pub column: Option<ColumnType>,
+    /// The file name of the source file.
+    pub file: Option<String>,
+    /// The directory of the source file.
+    #[serde(serialize_with = "serialize_typed_path")]
+    pub directory: Option<TypedPathBuf>,
+}
+
+impl SourceLocation {
+    /// Resolve debug information for a [`Instruction`] and create a [`SourceLocation`].
+    fn from_instruction(
+        debug_info: &DebugInfo,
+        program_unit: &unit_info::UnitInfo,
+        instruction: &Instruction,
+    ) -> Option<SourceLocation> {
+        debug_info
+            .find_file_and_directory(&program_unit.unit, instruction.file_index)
+            .map(|(file, directory)| SourceLocation {
+                line: instruction.line.map(std::num::NonZeroU64::get),
+                column: Some(instruction.column),
+                file,
+                directory,
+            })
+    }
+
+    /// Get the full path of the source file
+    pub fn combined_typed_path(&self) -> Option<TypedPathBuf> {
+        let combined_path = self
+            .directory
+            .as_ref()
+            .and_then(|dir| self.file.as_ref().map(|file| dir.join(file)));
+
+        combined_path
+    }
+}
