+++ probe-rs/src/debug/instructions/sequence.rs
@@ -1,0 +1,230 @@
+use super::{
+    super::{unit_info::UnitInfo, DebugError, DebugInfo},
+    block::Block,
+};
+use gimli::LineSequence;
+use std::{
+    self,
+    fmt::{Debug, Formatter},
+    ops::Range,
+};
+use typed_path::TypedPathBuf;
+
+/// Keep track of all the instruction locations required to satisfy the operations of [`SteppingMode`].
+/// This is a list of target instructions, belonging to a [`gimli::LineSequence`],
+/// and filters it to only user code instructions (no prologue code, and no non-statement instructions),
+/// so that we are left only with what DWARF terms as "recommended breakpoint location".
+pub(crate) struct Sequence<'debug_info> {
+    /// The `address_range.start` is the starting address of the program counter for which this sequence is valid,
+    /// and allows us to identify target instruction locations where the program counter lies inside the prologue.
+    /// The `address_range.end` is the first address that is not covered by this sequence within the line number program,
+    /// and allows us to identify when stepping over a instruction location would result in leaving a sequence.
+    /// - This is typically the instruction address of the first instruction in the next sequence,
+    ///   which may also be the first instruction in a new function.
+    pub(crate) address_range: Range<u64>,
+    /// See [`Block`].
+    pub(crate) blocks: Vec<Block>,
+    // The following private fields are required to resolve the source location information for
+    // each instruction location.
+    pub(crate) debug_info: &'debug_info DebugInfo,
+    pub(crate) program_unit: &'debug_info UnitInfo,
+}
+
+impl Debug for Sequence<'_> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        writeln!(
+            f,
+            "Sequence range: {:#010x}..{:#010x}",
+            self.address_range.start, self.address_range.end
+        )?;
+        for block in &self.blocks {
+            writeln!(
+                f,
+                "  Block range: {:#010x}..={:#010x}",
+                block.included_addresses.start(),
+                block.included_addresses.end()
+            )?;
+            for instruction in &block.instructions {
+                writeln!(
+                    f,
+                    "    {instruction:?} - {:?}",
+                    self.debug_info
+                        .get_path(&self.program_unit.unit, instruction.file_index)
+                        .map(
+                            |file_path| TypedPathBuf::from_unix(file_path.file_name().unwrap())
+                                .to_string_lossy()
+                                .to_string()
+                        )
+                        .unwrap_or("<unknown file>".to_string())
+                )?;
+            }
+            writeln!(f)?;
+        }
+        Ok(())
+    }
+}
+
+impl<'debug_info> Sequence<'debug_info> {
+    /// Extract all the instruction locations, belonging to the active sequence (i.e. the sequence that contains the `address`).
+    pub(crate) fn from_address(
+        debug_info: &'debug_info DebugInfo,
+        program_counter: u64,
+    ) -> Result<Self, DebugError> {
+        let program_unit = debug_info.compile_unit_info(program_counter)?;
+        let (offset, address_size) = if let Some(line_program) =
+            program_unit.unit.line_program.clone()
+        {
+            (
+                line_program.header().offset(),
+                line_program.header().address_size(),
+            )
+        } else {
+            let message = "The specified source location does not have any line_program information available. Please consider using instruction level stepping.".to_string();
+            return Err(DebugError::WarnAndContinue { message });
+        };
+
+        // Get the sequences of rows from the CompleteLineProgram at the given program_counter.
+        let incomplete_line_program =
+            debug_info
+                .debug_line_section
+                .program(offset, address_size, None, None)?;
+        let (complete_line_program, line_sequences) = incomplete_line_program.sequences()?;
+
+        // Get the sequence of rows that belongs to the program_counter.
+        let Some(line_sequence) = line_sequences.iter().find(|line_sequence| {
+            line_sequence.start <= program_counter && program_counter < line_sequence.end
+        }) else {
+            let message = "The specified source location does not have any line information available. Please consider using instruction level stepping.".to_string();
+            return Err(DebugError::WarnAndContinue { message });
+        };
+        let sequence = Self::from_line_sequence(
+            debug_info,
+            program_unit,
+            complete_line_program,
+            line_sequence,
+        );
+
+        if sequence.len() == 0 {
+            let message = "Could not find valid instruction locations for this address. Consider using instruction level stepping.".to_string();
+            Err(DebugError::WarnAndContinue { message })
+        } else {
+            tracing::trace!(
+                "Instruction location for pc={:#010x}\n{:?}",
+                program_counter,
+                sequence
+            );
+            Ok(sequence)
+        }
+    }
+
+    /// Build [`Sequence`] from a [`gimli::LineSequence`], with all the markers we need to determine valid halt locations.
+    pub(crate) fn from_line_sequence(
+        debug_info: &'debug_info DebugInfo,
+        program_unit: &'debug_info UnitInfo,
+        complete_line_program: gimli::CompleteLineProgram<
+            gimli::EndianReader<gimli::LittleEndian, std::rc::Rc<[u8]>>,
+            usize,
+        >,
+        line_sequence: &LineSequence<gimli::EndianReader<gimli::LittleEndian, std::rc::Rc<[u8]>>>,
+    ) -> Self {
+        let program_language = program_unit.get_language();
+        let mut sequence_rows = complete_line_program.resume_from(line_sequence);
+
+        // We have enough information to create the Sequence.
+        let mut sequence = Sequence {
+            address_range: line_sequence.start..line_sequence.end,
+            blocks: Vec::new(),
+            debug_info,
+            program_unit,
+        };
+
+        let mut block = Block {
+            included_addresses: line_sequence.start..=line_sequence.start,
+            instructions: Vec::new(),
+        };
+        let mut prologue_completed = false;
+        let mut previous_row: Option<gimli::LineRow> = None;
+
+        let sequence_functions = program_unit
+            .get_function_dies(debug_info, sequence.address_range.start + 8, true)
+            .ok();
+        if let Some(functions) = sequence_functions {
+            for function in functions {
+                let name = function.function_name(debug_info);
+                println!(
+                    "Function: {:?} @ {:#010x}:{:#010x}",
+                    name, function.low_pc, function.high_pc
+                );
+                // if let Some(ranges) = function.get_ranges(debug_info) {
+                //     for range in ranges {
+                //         if range.contains(&sequence.address_range.start) {
+                //             block.included_addresses = range.start..=range.end;
+                //         }
+                //     }
+                // }
+            }
+        }
+        while let Ok(Some((_, row))) = sequence_rows.next_row() {
+            if !prologue_completed && is_prologue_complete(row, program_language, previous_row) {
+                // This is the first row after the prologue, so we close off the prologue block, ...
+                sequence.blocks.push(block);
+                // ... and start a new block.
+                block = Block {
+                    included_addresses: row.address()..=row.address(),
+                    instructions: Vec::new(),
+                };
+                prologue_completed = true;
+            }
+
+            // The end of the sequence is not a valid halt location,
+            // nor is it a valid instruction in the current sequence.
+            if row.end_sequence() {
+                break;
+            }
+
+            // Logic to break the sequence into blocks. See [`Block`] for more information.
+            block.add(prologue_completed, row, previous_row.as_ref());
+            previous_row = Some(*row);
+        }
+        // Add the last block to the sequence.
+        if !block.instructions.is_empty() {
+            sequence.blocks.push(block);
+        }
+        sequence
+    }
+
+    /// Get the number of instruction locations in the list.
+    pub(crate) fn len(&self) -> usize {
+        self.blocks.len()
+    }
+}
+
+/// Test if the current row signals that we are beyond the prologue, and into user code
+pub(crate) fn is_prologue_complete(
+    row: &gimli::LineRow,
+    program_language: gimli::DwLang,
+    previous_row: Option<gimli::LineRow>,
+) -> bool {
+    let mut prologue_completed = row.prologue_end();
+
+    // For GNU C, it is known that the `DW_LNS_set_prologue_end` is not set, so we employ the same heuristic as GDB to determine when the prologue is complete.
+    // For other C compilers in the C99/11/17 standard, they will either set the `DW_LNS_set_prologue_end` or they will trigger this heuristic also.
+    // See https://gcc.gnu.org/legacy-ml/gcc-patches/2011-03/msg02106.html
+    if !prologue_completed
+        && matches!(
+            program_language,
+            gimli::DW_LANG_C99 | gimli::DW_LANG_C11 | gimli::DW_LANG_C17
+        )
+    {
+        if let Some(prev_row) = previous_row {
+            if row.end_sequence()
+                || (row.is_stmt()
+                    && (row.file_index() == prev_row.file_index()
+                        && (row.line() != prev_row.line() || row.line().is_none())))
+            {
+                prologue_completed = true;
+            }
+        }
+    }
+    prologue_completed
+}
